streamlit/requirements.txt
requests==2.27.1
pandas==1.4.2
numpy==1.22.4
streamlit==1.10.0
matplotlib
requests==2.26.0
fathon==1.3.1
scipy==1.8.0
gunicorn==20.1.0
google-cloud-storage
pandas_gbq == 0.17.1
pyarrow == 6.0.1
seaborn == 0.11.2
streamlit == 1.10.0
